requirements.txt
opencv-python
PyYAML
yamlordereddictloader
color_transfer